Move constants to a better place
