+++ custom/icds/const.py
@@ -5,3 +5,20 @@ CHILDREN_WEIGHED_REPORT_ID = 'static-icds-cas-static-ls_children_weighed'
 DAYS_AWC_OPEN_REPORT_ID = 'static-icds-cas-static-ls_awc_days_open'
 HOME_VISIT_REPORT_ID = 'static-icds-cas-static-ls_timely_home_visits'
 THR_REPORT_ID = 'static-icds-cas-static-ls_thr_30_days'
+
+AWC_LOCATION_TYPE_CODE = 'awc'
+SUPERVISOR_LOCATION_TYPE_CODE = 'supervisor'
+
+ANDHRA_PRADESH_SITE_CODE = '28'
+MAHARASHTRA_SITE_CODE = ''
+MADHYA_PRADESH_SITE_CODE = '23'
+BIHAR_SITE_CODE = '10'
+CHHATTISGARH_SITE_CODE = '22'
+JHARKHAND_SITE_CODE = '20'
+RAJASTHAN_SITE_CODE = '08'
+UTTAR_PRADESH_SITE_CODE = 'state9'
+
+ENGLISH = 'en'
+HINDI = 'hin'
+TELUGU = 'tel'
+MARATHI = 'mar'
